+++ src/net/l1j/server/datatables/UBTable.java
@@ -60,7 +60,7 @@ public class UBTable {
 			while (rs.next()) {
 				L1UltimateBattle ub = new L1UltimateBattle();
 				ub.setUbId(rs.getInt("ub_id"));
-				ub.setMapId(rs.getShort("ub_mapid"));
+				ub.setMapId(rs.getInt("ub_mapid"));
 				ub.setLocX1(rs.getInt("ub_area_x1"));
 				ub.setLocY1(rs.getInt("ub_area_y1"));
 				ub.setLocX2(rs.getInt("ub_area_x2"));
