+++ src/net/l1j/server/model/DungeonRandom.java
@@ -67,22 +67,22 @@ public class DungeonRandom {
 				String key = new StringBuilder().append(srcMapId).append(srcX).append(srcY).toString();
 				int[] newX = new int[5];
 				int[] newY = new int[5];
-				short[] newMapId = new short[5];
+				int[] newMapId = new int[5];
 				newX[0] = rs.getInt("new_x1");
 				newY[0] = rs.getInt("new_y1");
-				newMapId[0] = rs.getShort("new_mapid1");
+				newMapId[0] = rs.getInt("new_mapid1");
 				newX[1] = rs.getInt("new_x2");
 				newY[1] = rs.getInt("new_y2");
-				newMapId[1] = rs.getShort("new_mapid2");
+				newMapId[1] = rs.getInt("new_mapid2");
 				newX[2] = rs.getInt("new_x3");
 				newY[2] = rs.getInt("new_y3");
-				newMapId[2] = rs.getShort("new_mapid3");
+				newMapId[2] = rs.getInt("new_mapid3");
 				newX[3] = rs.getInt("new_x4");
 				newY[3] = rs.getInt("new_y4");
-				newMapId[3] = rs.getShort("new_mapid4");
+				newMapId[3] = rs.getInt("new_mapid4");
 				newX[4] = rs.getInt("new_x5");
 				newY[4] = rs.getInt("new_y5");
-				newMapId[4] = rs.getShort("new_mapid5");
+				newMapId[4] = rs.getInt("new_mapid5");
 				int heading = rs.getInt("new_heading");
 				NewDungeonRandom newDungeonRandom = new NewDungeonRandom(newX, newY, newMapId, heading);
 				if (_dungeonMap.containsKey(key)) {
@@ -100,10 +100,10 @@ public class DungeonRandom {
 	private static class NewDungeonRandom {
 		int[] _newX = new int[5];
 		int[] _newY = new int[5];
-		short[] _newMapId = new short[5];
+		int[] _newMapId = new int[5];
 		int _heading;
 
-		private NewDungeonRandom(int[] newX, int[] newY, short[] newMapId, int heading) {
+		private NewDungeonRandom(int[] newX, int[] newY, int[] newMapId, int heading) {
 			for (int i = 0; i < 5; i++) {
 				_newX[i] = newX[i];
 				_newY[i] = newY[i];
@@ -118,7 +118,7 @@ public class DungeonRandom {
 		if (_dungeonMap.containsKey(key)) {
 			int rnd = RandomArrayList.getInt(5);
 			NewDungeonRandom newDungeonRandom = _dungeonMap.get(key);
-			short newMap = newDungeonRandom._newMapId[rnd];
+			int newMap = newDungeonRandom._newMapId[rnd];
 			int newX = newDungeonRandom._newX[rnd];
 			int newY = newDungeonRandom._newY[rnd];
 			int heading = newDungeonRandom._heading;
